Add custom Python packaging for model bundles
Description and long description in bundle json files are not propagated
to the generated Python package. Previously, all model bundle packages
were generated with the same (statically defined) description and long-
description strings.  In addition, version and description are made
mandatory in the bundle json file.  If long description is left
unspecified, an empty long-description is used in the Python package.
Fixes issue #404.