fix memory leak errors in ydk_core_test
This commit also changed statements in test cases that cast raw pointer
to shared pointer. When a std::shared_ptr constructor is called with a
raw pointer, it creates a control block, and this control block will mess
up with original control block, which, with this commit, was added memory
cleanup functions.